requirements.txt
fastapi
uvicorn[standard]
openai
werewolf-agent-build-sdk==0.0.4
jinja2
langchain
httpx
requests